Refact: oauth login
