+++ src/main/java/com/WalkiePaw/domain/chatroom/service/ChatroomService.java
@@ -1,6 +1,7 @@
 package com.WalkiePaw.domain.chatroom.service;
 
 import com.WalkiePaw.domain.board.entity.Board;
+import com.WalkiePaw.domain.board.repository.BoardRepository;
 import com.WalkiePaw.domain.board.repository.BoardRepositoryOverride;
 import com.WalkiePaw.domain.chatroom.entity.Chatroom;
 import com.WalkiePaw.domain.chatroom.repository.ChatroomRepository;
@@ -22,7 +23,7 @@ import java.util.List;
 @Transactional(readOnly = true)
 public class ChatroomService {
     private final ChatroomRepository chatroomRepository;
-    private final BoardRepositoryOverride boardRepository;
+    private final BoardRepository boardRepository;
     private final MemberRepository memberRepository;
 
     public List<ChatroomListResponse> findAllByMemberId(final Integer memberId) {
