+++ src/main/java/com/WalkiePaw/presentation/domain/oauth/dto/OAuthUserinfoResponse.java
@@ -1,4 +1,4 @@
-package com.WalkiePaw.presentation.domain.oauth;
+package com.WalkiePaw.presentation.domain.oauth.dto;
 
 import lombok.Data;
