+++ src/main/java/com/WalkiePaw/presentation/domain/oauth/dto/OauthAccessToken.java
@@ -1,0 +1,26 @@
+package com.WalkiePaw.presentation.domain.oauth.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class OauthAccessToken {
+
+    @JsonProperty("access_token")
+    private String accessToken;
+
+    @JsonProperty("expires_in")
+    private int expiresIn;
+
+    @JsonProperty("token_in")
+    private int tokenIn;
+
+    @JsonProperty("scope")
+    private String scope;
+
+    @JsonProperty("refresh_token")
+    private String refreshToken;
+}
