+++ src/main/java/com/WalkiePaw/presentation/domain/oauth/provider/OAuthProvider.java
@@ -1,0 +1,7 @@
+package com.WalkiePaw.presentation.domain.oauth.provider;
+
+import com.WalkiePaw.presentation.domain.oauth.dto.OAuthUserinfoResponse;
+
+public interface OAuthProvider {
+    OAuthUserinfoResponse getUserInfo(final String code);
+}
